refactor: move `TaggedTable` into `safeds.data.tabular`
