--- src/safeds/data/__init__.py
@@ -1,1 +1,0 @@
-from ._tagged_table import TaggedTable
