+++ src/safeds/data/tabular/__init__.py
@@ -1,3 +1,4 @@
 from ._column import Column, ColumnStatistics
 from ._row import Row
 from ._table import Table
+from ._tagged_table import TaggedTable
