+++ src/safeds/ml/classification/_ada_boost.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 # noinspection PyProtectedMember
 import safeds.ml._util_sklearn
-from safeds.data import TaggedTable
+from safeds.data.tabular import TaggedTable
 from safeds.data.tabular import Table
 from sklearn.ensemble import AdaBoostClassifier as sk_AdaBoostClassifier
