+++ tests/safeds/data/tabular/_tagged_table/test_target_values.py
@@ -1,4 +1,4 @@
-from safeds.data import TaggedTable
+from safeds.data.tabular import TaggedTable
 from safeds.data.tabular import Table
