+++ tests/safeds/ml/classification/_k_nearest_neighbors/test_predict.py
@@ -1,5 +1,5 @@
 import pytest
-from safeds.data import TaggedTable
+from safeds.data.tabular import TaggedTable
 from safeds.data.tabular import Table
 from safeds.exceptions import PredictionError
 from safeds.ml.classification import KNearestNeighbors as KNearestNeighborsClassifier
