+++ tests/safeds/ml/regression/_elastic_net/test_fit.py
@@ -1,5 +1,5 @@
 import pytest
-from safeds.data import TaggedTable
+from safeds.data.tabular import TaggedTable
 from safeds.data.tabular import Table
 from safeds.exceptions import LearningError
 from safeds.ml.regression import ElasticNetRegression
