+++ tests/safeds/ml/test_util_sklearn.py
@@ -1,6 +1,6 @@
 import warnings
 
-from safeds.data import TaggedTable
+from safeds.data.tabular import TaggedTable
 from safeds.data.tabular import Table
 from safeds.ml.regression import LinearRegression
